[추가] 공고 등록 페이지 구현
